driver: initial version
The diagnostic driver provides its user with a possibility to direct
raw IPCs to the firmware.

This is done with reading and writing user-provided data to the HDA
PCI bus via IOCTLs interfacing with the device /dev/hda.

The driver can be built as follows:
    KDIR=<path to already built SOF Linux> make
which creates the file diag_driver.ko.

After loading the kernel module:
    insmod diag_driver.ko
the driver is ready to use.

The IOCTLs are:
    * CMD_OPEN_DEVICE          - opens the device,
    * CMD_ALLOC_MEMORY         - provides the host side with a memory
                                 buffer, i.e. allocates it, for HDA-DMA
                                 transaction between the host processor
                                 and the DSP,
    * CMD_FREE_MEMORY          - releases the memory allocated by
                                 CMD_ALLOC_MEMORY,
    * HDABUSTEST_READ_PCICONF  - reads the user-provided data from the PCI
                                 bus,
    * HDABUSTEST_WRITE_PCICONF - writes the user-provided data to the PCI
                                 bus.

Signed-off-by: Marcin Zielinski <marcinx.zielinski@intel.com>